# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Lab Launcher API",
    "version": "0.7.0",
    "description": "Pre-1.0 JSON API for the bundled frontend, trusted plugins, and private-deployment automation. Mutating API routes require the session CSRF token in the X-CSRF-Token header, except /api/bootstrap and /api/auth/login. Automation may instead authenticate with an Admin-issued API token via Authorization: Bearer; token requests skip CSRF, act with the token's role, and are rejected on session/profile endpoints (/api/auth/*, /api/me*, /api/bootstrap*)."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "security": [
    { "sessionCookie": [] },
    { "bearerToken": [] }
  ],
  "paths": {
...
    "/bootstrap": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create the first Admin account",
        "description": "Available only while no users exist. CSRF is not required for this bootstrap route.",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BootstrapRequest" } } } },
        "responses": { "200": { "description": "Created admin", "content": { "application/json": { "schema": { "type": "object", "required": ["user"], "properties": { "user": { "$ref": "#/components/schemas/User" }, "recoveryCodes": { "type": "array", "items": { "type": "string" }, "description": "Shown-once TOTP recovery codes when 2FA was enabled at bootstrap." } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Start a session",
        "description": "Returns a user and CSRF token, or requiresTotp=true after valid password authentication for a TOTP-enabled user. CSRF is not required for this login route.",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Login" }, "401": { "$ref": "#/components/responses/Error" }, "429": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "End the current session",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/me": {
      "get": {
        "tags": ["Profile"],
        "summary": "Read current user profile",
        "responses": { "200": { "$ref": "#/components/responses/UserEnvelope" }, "401": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/me/password": {
      "patch": {
        "tags": ["Profile"],
        "summary": "Change current user's password",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PasswordChangeRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/RevocationOk" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/me/sessions": {
      "get": {
        "tags": ["Profile"],
        "summary": "List current user's sessions",
        "responses": { "200": { "description": "Session list", "content": { "application/json": { "schema": { "type": "object", "required": ["sessions"], "properties": { "sessions": { "type": "array", "items": { "$ref": "#/components/schemas/UserSession" } } } } } } }, "401": { "$ref": "#/components/responses/Error" } }
      },
      "delete": {
        "tags": ["Profile"],
        "summary": "Revoke other sessions for current user",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "responses": { "200": { "description": "Revocation count", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Ok" }, { "type": "object", "properties": { "count": { "type": "integer" } } }] } } } }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/me/sessions/{sid}": {
      "delete": {
        "tags": ["Profile"],
        "summary": "Revoke one current-user session",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Sid" }],
        "responses": { "200": { "$ref": "#/components/responses/Ok" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/me/totp/setup": {
      "post": {
        "tags": ["Profile"],
        "summary": "Generate a TOTP setup secret",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "responses": { "200": { "description": "Generated secret", "content": { "application/json": { "schema": { "type": "object", "required": ["secret"], "properties": { "secret": { "type": "string" } } } } } }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/me/totp/enable": {
      "post": {
        "tags": ["Profile"],
        "summary": "Enable TOTP 2FA for current user",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TotpEnableRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/RecoveryCodesOk" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/me/totp/recovery-codes": {
      "post": {
        "tags": ["Profile"],
        "summary": "Regenerate TOTP recovery codes",
        "description": "Replaces all remaining recovery codes with a fresh shown-once set of ten single-use codes. Requires the current password and a valid TOTP code.",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["password", "code"], "properties": { "password": { "type": "string" }, "code": { "type": "string" } } } } } },
        "responses": { "200": { "$ref": "#/components/responses/RecoveryCodesOk" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/me/totp/disable": {
      "post": {
        "tags": ["Profile"],
        "summary": "Disable TOTP 2FA for current user",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TotpDisableRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Ok" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "hll.sid" },
      "csrfToken": { "type": "apiKey", "in": "header", "name": "X-CSRF-Token" },
      "bearerToken": { "type": "http", "scheme": "bearer", "description": "Admin-issued API token (hll_ prefix). Skips CSRF; not accepted on session/profile endpoints." }
    },
    "parameters": {
...
      "Sid": { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "Ok": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } },
      "RevocationOk": { "description": "Success with session revocation count", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RevocationOk" } } } },
      "Error": { "description": "Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Login": { "description": "Login response", "content": { "application/json": { "schema": { "oneOf": [{ "type": "object", "required": ["user", "csrfToken"], "properties": { "user": { "$ref": "#/components/schemas/User" }, "csrfToken": { "type": "string" } } }, { "type": "object", "required": ["requiresTotp"], "properties": { "requiresTotp": { "type": "boolean", "const": true } } }] } } } },
      "UserEnvelope": { "description": "User envelope", "content": { "application/json": { "schema": { "type": "object", "required": ["user"], "properties": { "user": { "$ref": "#/components/schemas/User" } } } } } },
...
      "RecoveryCodesOk": { "description": "Success with shown-once recovery codes", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Ok" }, { "type": "object", "required": ["recoveryCodes"], "properties": { "recoveryCodes": { "type": "array", "items": { "type": "string" }, "description": "Ten single-use codes, shown exactly once." } } }] } } } }
    },
    "schemas": {
      "Ok": { "type": "object", "required": ["ok"], "properties": { "ok": { "type": "boolean" } }, "additionalProperties": true },
      "RevocationOk": { "allOf": [{ "$ref": "#/components/schemas/Ok" }, { "type": "object", "properties": { "revokedSessions": { "type": "integer", "minimum": 0 } } }] },
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } }, "additionalProperties": true },
...
      "Role": { "type": "string", "enum": ["admin", "editor", "user"] },
      "User": { "type": "object", "properties": { "id": { "type": "integer" }, "username": { "type": "string" }, "role": { "$ref": "#/components/schemas/Role" }, "totpEnabled": { "type": ["boolean", "integer"] }, "recoveryCodesRemaining": { "type": "integer", "description": "Unused TOTP recovery codes. Present on /api/me when 2FA is enabled." }, "createdAt": { "type": "string" }, "updatedAt": { "type": "string" } }, "additionalProperties": true },
      "BootstrapRequest": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string", "minLength": 3 }, "password": { "type": "string", "minLength": 10 }, "totpSecret": { "type": "string" }, "totpCode": { "type": "string" } } },
      "LoginRequest": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string" }, "password": { "type": "string" }, "code": { "type": "string", "description": "Six-digit TOTP code." }, "recoveryCode": { "type": "string", "description": "Single-use recovery code accepted in place of a TOTP code; case and separators are ignored." } } },
...
      "PasswordChangeRequest": { "type": "object", "required": ["currentPassword", "newPassword"], "properties": { "currentPassword": { "type": "string" }, "newPassword": { "type": "string", "minLength": 10 } } },
      "TotpEnableRequest": { "type": "object", "required": ["secret", "code"], "properties": { "secret": { "type": "string" }, "code": { "type": "string" } } },
      "TotpDisableRequest": { "type": "object", "required": ["password"], "properties": { "password": { "type": "string" }, "code": { "type": "string" } } },
      "UserSession": { "type": "object", "properties": { "sid": { "type": "string" }, "userAgent": { "type": "string" }, "ip": { "type": "string" }, "createdAt": { "type": ["string", "null"] }, "expiresAt": { "type": "string" }, "current": { "type": "boolean" } }, "additionalProperties": true },
...
    }
  }
}